Adicionada exceção para DB 'foreign key constraint', adicionado DELETE/CRUD.

+++ src/main/java/com/dutra/dscomerce/services/ProductService.java
@@ -4,11 +4,15 @@ import com.dutra.dscomerce.dtos.ProductDto;
 import com.dutra.dscomerce.dtos.ProductEntry;
 import com.dutra.dscomerce.entities.ProductEntity;
 import com.dutra.dscomerce.repositories.ProductRepository;
+import com.dutra.dscomerce.services.exceptions.DatabaseException;
 import com.dutra.dscomerce.services.exceptions.ResourceNotFoundException;
 import com.dutra.dscomerce.services.interfaces.ProductServiceInterface;
+import jakarta.persistence.EntityNotFoundException;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
@@ -42,20 +46,34 @@ public class ProductService implements ProductServiceInterface {
     @Override
     @Transactional
     public ProductDto updateProduct(Long id, ProductEntry product) {
-        ProductEntity updatedProduct = repository.getReferenceById(id);
+        try {
+            ProductEntity updatedProduct = repository.getReferenceById(id);
 
-        updatedProduct.setImgUrl(product.imgUrl());
-        updatedProduct.setPrice(product.price());
-        updatedProduct.setName(product.name());
-        updatedProduct.setDescription(product.description());
+            updatedProduct.setImgUrl(product.imgUrl());
+            updatedProduct.setPrice(product.price());
+            updatedProduct.setName(product.name());
+            updatedProduct.setDescription(product.description());
 
-        return builderDto(repository.save(updatedProduct));
+            return builderDto(repository.save(updatedProduct));
+        } catch (EntityNotFoundException exception) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
     }
 
     @Override
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
-        repository.deleteById(id);
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
+        try {
+            repository.deleteById(id);
+        }
+        catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Violação de integridade referencial / FK");
+        }
     }
+
 
     private ProductDto builderDto(ProductEntity product) {
         return new ProductDto(product.getId(),
